rename acs1 to methodfee
