+++ src/AElf.CrossChain.Core/CrossChainContractChargeFeeStrategy.cs
@@ -1,6 +1,6 @@
 using System.Linq;
 using AElf.Kernel.SmartContract.Application;
-using AElf.Kernel.SmartContract.ExecutionPluginForAcs1.FreeFeeTransactions;
+using AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.FreeFeeTransactions;
 using AElf.Types;
 
 namespace AElf.CrossChain
