+++ src/AElf.Kernel.Consensus.AEDPoS/ConsensusContractChargeFeeStrategy.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using AElf.Contracts.Consensus.AEDPoS;
 using AElf.Kernel.SmartContract.Application;
-using AElf.Kernel.SmartContract.ExecutionPluginForAcs1.FreeFeeTransactions;
+using AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.FreeFeeTransactions;
 using AElf.Types;
 
 namespace AElf.Kernel.Consensus.AEDPoS
