+++ src/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.csproj
@@ -3,7 +3,7 @@
 
     <PropertyGroup>
         <TargetFramework>netstandard2.0</TargetFramework>
-        <PackageId>AElf.Kernel.SmartContract.ExecutionPluginForAcs1</PackageId>
+        <PackageId>AElf.Kernel.SmartContract.ExecutionPluginForMethodFee</PackageId>
         <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
         <Description>ACS 1 execution plugin.</Description>
     </PropertyGroup>
