+++ src/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee/FreeFeeTransactions/TokenContractChargeFeeStrategy.cs
@@ -1,11 +1,10 @@
 using System.Collections.Generic;
 using AElf.Contracts.MultiToken;
 using AElf.Kernel.SmartContract.Application;
-using AElf.Kernel.SmartContract.ExecutionPluginForAcs1.FreeFeeTransactions;
 using AElf.Kernel.Token;
 using AElf.Types;
 
-namespace AElf.Kernel.SmartContract.ExecutionPluginForAcs1.Tests
+namespace AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.FreeFeeTransactions
 {
     public class TokenContractChargeFeeStrategy : IChargeFeeStrategy
     {
@@ -23,10 +22,20 @@ namespace AElf.Kernel.SmartContract.ExecutionPluginForAcs1.Tests
 
         public bool IsFree(Transaction transaction)
         {
+            // Stop charging fee from system txs and plugin txs.
             return new List<string>
             {
-                nameof(TokenContractContainer.TokenContractStub.Create),
-                nameof(TokenContractContainer.TokenContractStub.Issue),
+                // System tx
+                nameof(TokenContractContainer.TokenContractStub.ClaimTransactionFees),
+                nameof(TokenContractContainer.TokenContractStub.DonateResourceToken),
+
+                // Pre-plugin tx
+                nameof(TokenContractContainer.TokenContractStub.ChargeTransactionFees),
+                nameof(TokenContractContainer.TokenContractStub.CheckThreshold),
+                nameof(TokenContractContainer.TokenContractStub.CheckResourceToken),
+
+                // Post-plugin tx
+                nameof(TokenContractContainer.TokenContractStub.ChargeResourceToken),
             }.Contains(transaction.MethodName);
         }
     }
