+++ src/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee/MethodStubFactory.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using AElf.CSharp.Core;
 using Google.Protobuf;
 
-namespace AElf.Kernel.SmartContract.ExecutionPluginForAcs1
+namespace AElf.Kernel.SmartContract.ExecutionPluginForMethodFee
 {
     //TODO: make a base class
     public class MethodStubFactory : IMethodStubFactory
