+++ src/AElf.Kernel.SmartContract.ExecutionPluginForProposal/AElf.Kernel.SmartContract.ExecutionPluginForProposal.csproj
@@ -9,10 +9,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\AElf.Kernel.SmartContract.ExecutionPluginForAcs1\AElf.Kernel.SmartContract.ExecutionPluginForAcs1.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
         <ContractStub Include="..\..\protobuf\acs1.proto">
             <Link>Protobuf\Proto\acs1.proto</Link>
         </ContractStub>
@@ -26,6 +22,10 @@
 
     <ItemGroup>
       <PackageReference Include="Volo.Abp.Core" Version="1.1.2" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\AElf.Kernel.SmartContract.ExecutionPluginForMethodFee\AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.csproj" />
     </ItemGroup>
 
 </Project>
