+++ src/AElf.Kernel.SmartContract.ExecutionPluginForProposal/ParliamentContractChargeFeeStrategy.cs
@@ -1,6 +1,6 @@
 using AElf.Contracts.Parliament;
 using AElf.Kernel.SmartContract.Application;
-using AElf.Kernel.SmartContract.ExecutionPluginForAcs1.FreeFeeTransactions;
+using AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.FreeFeeTransactions;
 using AElf.Types;
 
 namespace AElf.Kernel.SmartContract.ExecutionPluginForProposal
