+++ src/AElf.Kernel.SmartContractExecution/SmartContractExecutionAElfModule.cs
@@ -1,6 +1,6 @@
 ﻿using AElf.Kernel.SmartContract;
 using AElf.Kernel.SmartContract.Application;
-using AElf.Kernel.SmartContract.ExecutionPluginForAcs1;
+using AElf.Kernel.SmartContract.ExecutionPluginForMethodFee;
 using AElf.Kernel.SmartContract.ExecutionPluginForAcs5;
 using AElf.Kernel.SmartContract.ExecutionPluginForAcs8;
 using AElf.Kernel.SmartContract.ExecutionPluginForProposal;
@@ -16,7 +16,7 @@ namespace AElf.Kernel.SmartContractExecution
         typeof(ExecutionPluginForAcs8Module),
         typeof(ExecutionPluginForAcs5Module),
         typeof(ExecutionPluginForProposalModule),
-        typeof(ExecutionPluginForAcs1Module))]
+        typeof(ExecutionPluginForMethodFeeModule))]
     public class SmartContractExecutionAElfModule : AElfModule
     {
         public override void ConfigureServices(ServiceConfigurationContext context)
