+++ test/AElf.Contracts.TestContract.Tests/TestContractAElfModule.cs
@@ -1,8 +1,8 @@
 using AElf.Contracts.TestKit;
 using AElf.Kernel.SmartContract;
 using AElf.Kernel.SmartContract.Application;
-using AElf.Kernel.SmartContract.ExecutionPluginForAcs1;
-using AElf.Kernel.SmartContract.ExecutionPluginForAcs1.FreeFeeTransactions;
+using AElf.Kernel.SmartContract.ExecutionPluginForMethodFee;
+using AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.FreeFeeTransactions;
 using AElf.Kernel.SmartContract.ExecutionPluginForAcs8;
 using Microsoft.Extensions.DependencyInjection;
 using Volo.Abp.Modularity;
@@ -20,7 +20,7 @@ namespace AElf.Contract.TestContract
 
     [DependsOn(
         typeof(ContractTestModule),
-        typeof(ExecutionPluginForAcs1Module),
+        typeof(ExecutionPluginForMethodFeeModule),
         typeof(ExecutionPluginForAcs8Module))]
     public class TestFeesContractAElfModule : ContractTestModule
     {
