--- test/AElf.Kernel.SmartContract.ExecutionPluginForAcs1.Tests/ExecutionPluginForAcs1TestBase.cs
@@ -1,8 +1,0 @@
-using AElf.Contracts.TestKit;
-
-namespace AElf.Kernel.SmartContract.ExecutionPluginForAcs1.Tests
-{
-    public class ExecutionPluginForAcs1TestBase : ContractTestBase<ExecutionPluginForAcs1TestModule>
-    {
-    }
-}
