+++ test/AElf.Kernel.SmartContract.ExecutionPluginForAcs5.Tests/AElf.Kernel.SmartContract.ExecutionPluginForAcs5.Tests.csproj
@@ -20,7 +20,7 @@
         <ProjectReference Include="..\..\src\AElf.Kernel.SmartContract.ExecutionPluginForAcs5\AElf.Kernel.SmartContract.ExecutionPluginForAcs5.csproj" />
     </ItemGroup>
     <ItemGroup>
-        <ProjectReference Include="..\AElf.Kernel.SmartContract.ExecutionPluginForAcs1.Tests\AElf.Kernel.SmartContract.ExecutionPluginForAcs1.Tests.csproj" />
+        <ProjectReference Include="..\AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests\AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests.csproj" />
         <ProjectReference Include="..\AElf.Kernel.SmartContract.ExecutionPluginForAcs5.Tests.TestContract\AElf.Kernel.SmartContract.ExecutionPluginForAcs5.Tests.TestContract.csproj">
             <ReferenceOutputAssembly>false</ReferenceOutputAssembly>
             <OutputItemType>Contract</OutputItemType>
