+++ test/AElf.Kernel.SmartContract.ExecutionPluginForAcs8.Tests/ExecutionPluginForAcs8TestModule.cs
@@ -1,6 +1,6 @@
 using AElf.Contracts.TestKit;
 using AElf.Kernel.SmartContract.Application;
-using AElf.Kernel.SmartContract.ExecutionPluginForAcs1.FreeFeeTransactions;
+using AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.FreeFeeTransactions;
 using Microsoft.Extensions.DependencyInjection;
 using Volo.Abp.Modularity;
