+++ test/AElf.Kernel.SmartContract.ExecutionPluginForAcs8.Tests/TokenContractChargeFeeStrategy.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using AElf.Contracts.MultiToken;
 using AElf.Kernel.SmartContract.Application;
-using AElf.Kernel.SmartContract.ExecutionPluginForAcs1.FreeFeeTransactions;
+using AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.FreeFeeTransactions;
 using AElf.Kernel.Token;
 using AElf.Types;
