+++ test/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests.TestContract/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests.TestContract.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
         <TargetFramework>netstandard2.0</TargetFramework>
-        <PackageId>AElf.Kernel.SmartContract.ExecutionPluginForAcs1.Tests.TestContract</PackageId>
+        <PackageId>AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests.TestContract</PackageId>
         <IsContract>true</IsContract>
     </PropertyGroup>
     <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
