+++ test/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests.TestContract/ContractState.cs
@@ -1,7 +1,7 @@
 using Acs1;
 using AElf.Sdk.CSharp.State;
 
-namespace AElf.Kernel.SmartContract.ExecutionPluginForAcs1.Tests.TestContract
+namespace AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests.TestContract
 {
     public class ContractState : AElf.Sdk.CSharp.State.ContractState
     {
