+++ test/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests/ExecutionPluginForMethodFeeTestBase.cs
@@ -1,0 +1,8 @@
+using AElf.Contracts.TestKit;
+
+namespace AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests
+{
+    public class ExecutionPluginForMethodFeeTestBase : ContractTestBase<ExecutionPluginForMethodFeeTestModule>
+    {
+    }
+}
