+++ test/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests/ExecutionPluginForMethodFeeTestModule.cs
@@ -1,14 +1,14 @@
 using AElf.Contracts.TestKit;
 using AElf.Kernel.SmartContract.Application;
-using AElf.Kernel.SmartContract.ExecutionPluginForAcs1.FreeFeeTransactions;
+using AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.FreeFeeTransactions;
 using Microsoft.Extensions.DependencyInjection;
 using Volo.Abp.Modularity;
 
-namespace AElf.Kernel.SmartContract.ExecutionPluginForAcs1.Tests
+namespace AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests
 {
     [DependsOn(typeof(ContractTestModule),
-        typeof(ExecutionPluginForAcs1Module))]
-    public class ExecutionPluginForAcs1TestModule : ContractTestModule
+        typeof(ExecutionPluginForMethodFeeModule))]
+    public class ExecutionPluginForMethodFeeTestModule : ContractTestModule
     {
         public override void ConfigureServices(ServiceConfigurationContext context)
         {
