+++ test/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests/TestCalculateStrategy.cs
@@ -4,7 +4,7 @@ using AElf.Kernel.SmartContract.Application;
 using AElf.Kernel.TransactionPool.Application;
 using AElf.Kernel.TransactionPool.Infrastructure;
 
-namespace AElf.Kernel.SmartContract.ExecutionPluginForAcs1.Tests
+namespace AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests
 {
      public class TestCalculateTxStrategy : CalculateCostStrategyBase, ICalculateTxCostStrategy
     {
