+++ test/AElf.Parallel.Tests/ParallelTestAElfModule.cs
@@ -2,7 +2,7 @@ using AElf.Kernel.SmartContract.Application;
 using System.Collections.Generic;
 using AElf.Kernel;
 using AElf.Kernel.SmartContract;
-using AElf.Kernel.SmartContract.ExecutionPluginForAcs1;
+using AElf.Kernel.SmartContract.ExecutionPluginForMethodFee;
 using AElf.Kernel.SmartContract.Parallel;
 using AElf.Modularity;
 using AElf.OS;
